Fix incorrect ViewModel implementation
