+++ app/src/main/java/com/example/kotlinapp/App.kt
@@ -2,13 +2,10 @@ package com.example.kotlinapp
 
 import android.app.Application
 import androidx.room.Room
-import com.example.kotlinapp.pokemon_database.AppDatabase
-import com.example.kotlinapp.pokemon_info.PokemonInfoViewModel
-import com.example.kotlinapp.pokemon_list.PokemonListViewModel
-import com.example.kotlinapp.pokemon_network.PokemonNetwork
 
 class App : Application() {
-    val pokemonNetwork = PokemonNetwork()
+    val pokemonRepository = PokemonRepository()
+    lateinit var db: AppDatabase
 
     override fun onCreate() {
         super.onCreate()
@@ -22,8 +19,5 @@ class App : Application() {
 
     companion object {
         lateinit var instance: App
-        lateinit var db: AppDatabase
-        lateinit var pokemonListViewModel: PokemonListViewModel
-        lateinit var pokemonInfoViewModel: PokemonInfoViewModel
     }
 }
