+++ app/src/main/java/com/example/kotlinapp/PokemonInfoViewModelFactory.kt
@@ -1,14 +1,14 @@
-package com.example.kotlinapp.pokemon_info
+package com.example.kotlinapp
 
 import android.app.Application
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 
 class PokemonInfoViewModelFactory(
-    private val fragment: PokemonInfoFragment,
+    private val name: String,
     val app: Application
 ) : ViewModelProvider.AndroidViewModelFactory(app) {
     override fun <T : ViewModel> create(modelClass: Class<T>): T {
-        return PokemonInfoViewModel(fragment) as T
+        return PokemonInfoViewModel(name) as T
     }
 }
