+++ app/src/main/java/com/example/kotlinapp/PokemonListFragment.kt
@@ -1,0 +1,102 @@
+package com.example.kotlinapp
+
+import android.os.Bundle
+import android.util.Log
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import android.widget.Toast
+import androidx.fragment.app.Fragment
+import androidx.lifecycle.MutableLiveData
+import androidx.lifecycle.Observer
+import androidx.lifecycle.ViewModelProvider
+import androidx.recyclerview.widget.LinearLayoutManager
+import androidx.recyclerview.widget.RecyclerView
+import com.example.kotlinapp.databinding.FragmentMainBinding
+
+class PokemonListFragment : Fragment() {
+    private var _binding: FragmentMainBinding? = null
+    private val binding get() = _binding!!
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?
+    ): View {
+        _binding = FragmentMainBinding.inflate(inflater)
+        return binding.root
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        val pokemonListViewModel = ViewModelProvider(this)[PokemonListViewModel::class.java]
+        val pokemonListLiveData = pokemonListViewModel.pokemonItemListLiveData
+
+        val recyclerView: RecyclerView = binding.recyclerView
+        recyclerView.layoutManager = LinearLayoutManager(activity)
+
+        val adapter = PokemonAdapter(
+            onPokemonClick = { pokemonItem: PokemonItem ->
+                val bundle = Bundle()
+
+                val pokemon =
+                    pokemonListLiveData.value!!.firstOrNull { it.name == pokemonItem.name }
+
+                bundle.putString("pokemonName", pokemon!!.name)
+
+                parentFragmentManager.beginTransaction()
+                    .replace(R.id.fragment, PokemonInfoFragment::class.java, bundle)
+                    .addToBackStack("PokemonListFragment").commit()
+            },
+            onIsFavoriteClick = { pokemonItem: PokemonItem ->
+                pokemonListViewModel.toggleFavorite(pokemonItem)
+            }
+        )
+
+
+        pokemonListLiveData.observe(viewLifecycleOwner) {
+            adapter.setPokemons(it)
+        }
+        val nextPageLoadingState = pokemonListViewModel.nextPageLoadingState
+        nextPageLoadingState.observe(viewLifecycleOwner) {
+            with (nextPageLoadingState.value!!) {
+                if (isLoaded) {
+                    showSuccessMessage()
+                } else {
+                    handleNetworkError()
+                    Log.d(
+                        "next page loading failure",
+                        error!!.message.toString()
+                    )
+                }
+            }
+        }
+
+        recyclerView.adapter = adapter
+
+        recyclerView.addOnScrollListener(
+            object : RecyclerView.OnScrollListener() {
+                override fun onScrollStateChanged(recyclerView: RecyclerView, newState: Int) {
+                    super.onScrollStateChanged(recyclerView, newState)
+
+                    if (!recyclerView.canScrollVertically(1) &&
+                        newState == RecyclerView.SCROLL_STATE_IDLE
+                    ) {
+                        Toast.makeText(context, "загрузка...", Toast.LENGTH_LONG).show()
+
+                        pokemonListViewModel.loadNextPage()
+                    }
+                }
+            }
+        )
+    }
+
+    private fun handleNetworkError() {
+        requireActivity().runOnUiThread {
+            Toast.makeText(requireActivity(), "Ошибка сети", Toast.LENGTH_LONG).show()
+        }
+    }
+
+    private fun showSuccessMessage() {
+        requireActivity().runOnUiThread {
+            Toast.makeText(requireActivity(), "Загрузка завершена", Toast.LENGTH_LONG).show()
+        }
+    }
+}
